pyproject.toml
[tool.poetry]
name = "easyca"
version = "2.0.0"
description = "CA management and TLS certificates generation"
authors = ["Louis Cailliot <louis.cailliot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.1"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
easyca = "easyca.main:main"
